更新 AXI_Memory 和相关测试文件，主要改动包括：将 Memory_Item 中的 data 类型从 Long 修改为 UInt，调整 AXI_Read_Item 和 AXI_Write_Item 中的 Boolean 类型为 Scala 的 Boolean，优化内存读写逻辑以支持 UInt 类型，新增从文件加载内存数据的功能，并更新测试类以适应这些更改。

+++ src/test/scala/L2Cache_Tester.scala
@@ -1,4 +1,4 @@
-import chisel3._
+// import chisel3._
 import chiseltest._
 import org.scalatest.flatspec.AnyFlatSpec
 import scala.collection.mutable.Queue
@@ -6,6 +6,7 @@ import scala.util._
 import CPU_Config.Cache._
 import Zircon_Util._
 import os.write
+import spire.math.UInt
 
 object L2_Test_Config {
     val total_space             = 4096
@@ -115,7 +116,7 @@ class L2Cache_Tester extends AnyFlatSpec with ChiselScalatestTester{
                 c.io.axi.arready.poke(r.arready)
                 c.io.axi.awready.poke(w.awready)
                 c.io.axi.bvalid.poke(w.bvalid)
-                c.io.axi.rdata.poke(r.rdata)
+                c.io.axi.rdata.poke(r.rdata.toInt)
                 c.io.axi.rlast.poke(r.rlast)
                 c.io.axi.rvalid.poke(r.rvalid)
                 c.io.axi.wready.poke(w.wready)
